fix data resets and rename files
