+++ companion/src/generaledit/generalfavorites.h
