+++ companion/src/generaledit/generalkeyshortcuts.h
